CPLP-1458-controller-cleanup
