+++ src/marketplace/Apps.Service/BusinessLogic/IAppReleaseBusinessLogic.cs
@@ -22,6 +22,7 @@ using Org.CatenaX.Ng.Portal.Backend.Apps.Service.ViewModels;
 using Org.CatenaX.Ng.Portal.Backend.PortalBackend.PortalEntities.Enums;
 using Org.CatenaX.Ng.Portal.Backend.PortalBackend.DBAccess.Models;
 using Org.CatenaX.Ng.Portal.Backend.Offers.Library.Models;
+using Org.CatenaX.Ng.Portal.Backend.Framework.Models;
 
 namespace Org.CatenaX.Ng.Portal.Backend.Apps.Service.BusinessLogic;
 
@@ -98,4 +99,25 @@ public interface IAppReleaseBusinessLogic
     /// <param name="iamUserId"></param>
     /// <returns></returns>
     Task DeleteAppRoleAsync(Guid appId, Guid roleId, string iamUserId);
+
+    /// <summary>
+    /// Creates an application and returns its generated ID.
+    /// </summary>
+    /// <param name="appRequestModel"></param>
+    /// <returns>Guid of the created app.</returns>
+    Task<Guid> AddAppAsync(AppRequestModel appRequestModel);
+
+    /// <summary>
+    /// Retrieves all in review status apps in the marketplace.
+    /// </summary>
+    Task<Pagination.Response<InReviewAppData>> GetAllInReviewStatusAppsAsync(int page = 0, int size = 15);
+
+    /// <summary>
+    /// Update app status and create notification
+    /// </summary>
+    /// <param name="appId"></param>
+    /// <param name="iamUserId"></param>
+    /// <returns></returns>
+    Task SubmitAppReleaseRequestAsync(Guid appId, string iamUserId);
+
 }
