+++ src/marketplace/Apps.Service/Controllers/AppReleaseProcessController.cs
@@ -27,6 +27,7 @@ using Microsoft.AspNetCore.Authorization;
 using Org.CatenaX.Ng.Portal.Backend.PortalBackend.DBAccess.Models;
 using Microsoft.AspNetCore.Mvc;
 using Org.CatenaX.Ng.Portal.Backend.Offers.Library.Models;
+using Org.CatenaX.Ng.Portal.Backend.Framework.Models;
 
 namespace Org.CatenaX.Ng.Portal.Backend.Apps.Service.Controllers;
 
@@ -192,6 +193,61 @@ public class AppReleaseProcessController : ControllerBase
     public async Task<IActionResult> DeleteAppRoleAsync([FromRoute] Guid appId, [FromRoute] Guid roleId)
     {
         await this.WithIamUserId(iamUserId => _appReleaseBusinessLogic.DeleteAppRoleAsync(appId, roleId, iamUserId));
+        return NoContent();
+    }
+
+    /// <summary>
+    /// Creates an app according to request model
+    /// </summary>
+    /// <param name="appRequestModel">Request model for app creation.</param>
+    /// <returns>ID of created application.</returns> 
+    /// <remarks>Example: POST: /api/apps/appreleaseprocess/createapp</remarks>
+    /// <response code="201">Returns created app's ID.</response>
+    /// <response code="404">Language Code or Use Case or CompanyId does not exist.</response>
+    [HttpPost]
+    [Route("createapp")]
+    [Authorize(Roles = "add_app")]
+    [ProducesResponseType(typeof(Guid), StatusCodes.Status201Created)]
+    [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(typeof(ErrorResponse),StatusCodes.Status404NotFound)]
+    public async Task<ActionResult<Guid>> ExecuteAppCreation([FromBody] AppRequestModel appRequestModel)
+    {
+        var appId = await _appReleaseBusinessLogic.AddAppAsync(appRequestModel).ConfigureAwait(false);
+        return CreatedAtRoute("GetAppDetailsByIdAsync", new {controller = "Apps",appId = appId}, appId);
+    }
+
+    /// <summary>
+    /// Retrieves all in review status apps in the marketplace .
+    /// </summary>
+    /// <param name="page">page index start from 0</param>
+    /// <param name="size">size to get number of records</param>
+    /// <returns>Collection of all in review status marketplace apps.</returns>
+    /// <remarks>Example: GET: /api/apps/appreleaseprocess/inReview</remarks>
+    /// <response code="200">Returns the list of all in review status marketplace apps.</response>
+    [HttpGet]
+    [Route("inReview")]
+    [Authorize(Roles = "approve_app_release,decline_app_release")]
+    [ProducesResponseType(typeof(Pagination.Response<InReviewAppData>), StatusCodes.Status200OK)]
+    public Task<Pagination.Response<InReviewAppData>> GetAllInReviewStatusAppsAsync([FromQuery] int page = 0,[FromQuery] int size = 15) =>
+        _appReleaseBusinessLogic.GetAllInReviewStatusAppsAsync( page,size);
+
+    /// <summary>
+    /// Submit an app for release
+    /// </summary>
+    /// <param name="appId" example="D3B1ECA2-6148-4008-9E6C-C1C2AEA5C645">ID of the app.</param>
+    /// <remarks>Example: PUT: /api/apps/appreleaseprocess/D3B1ECA2-6148-4008-9E6C-C1C2AEA5C645/submit</remarks>
+    /// <response code="204">The app was successfully submitted for release.</response>
+    /// <response code="400">Either the sub claim is empty/invalid, user does not exist or the subscription might not have the correct status or the companyID is incorrect.</response>
+    /// <response code="404">App does not exist.</response>
+    [HttpPut]
+    [Route("{appId}/submit")]
+    [Authorize(Roles = "add_app")]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
+    public async Task<IActionResult> SubmitAppReleaseRequest([FromRoute] Guid appId)
+    {
+        await this.WithIamUserId(userId => _appReleaseBusinessLogic.SubmitAppReleaseRequestAsync(appId, userId)).ConfigureAwait(false);
         return NoContent();
     }
 }
