+++ src/marketplace/Apps.Service/Controllers/AppsController.cs
@@ -21,7 +21,6 @@
 using Org.CatenaX.Ng.Portal.Backend.Apps.Service.BusinessLogic;
 using Org.CatenaX.Ng.Portal.Backend.Apps.Service.ViewModels;
 using Org.CatenaX.Ng.Portal.Backend.Framework.ErrorHandling;
-using Org.CatenaX.Ng.Portal.Backend.Framework.Models;
 using Org.CatenaX.Ng.Portal.Backend.Keycloak.Authentication;
 using Org.CatenaX.Ng.Portal.Backend.PortalBackend.DBAccess.Models;
 using Org.CatenaX.Ng.Portal.Backend.PortalBackend.PortalEntities.Enums;
@@ -269,62 +268,6 @@ public class AppsController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status400BadRequest)]
     public IAsyncEnumerable<AllAppData> GetAppDataAsync()=>
         this.WithIamUserId(userId => _appsBusinessLogic.GetCompanyProvidedAppsDataForUserAsync(userId));
-    
-
-    /// <summary>
-    /// Creates an app according to request model
-    /// </summary>
-    /// <param name="appRequestModel">Request model for app creation.</param>
-    /// <returns>ID of created application.</returns> 
-    /// <remarks>Example: POST: /api/apps/createapp</remarks>
-    /// <response code="201">Returns created app's ID.</response>
-    /// <response code="404">Language Code or Use Case or CompanyId does not exist.</response>
-    [HttpPost]
-    [Route("createapp")]
-    [Authorize(Roles = "add_app")]
-    [ProducesResponseType(typeof(Guid), StatusCodes.Status201Created)]
-    [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(typeof(ErrorResponse),StatusCodes.Status404NotFound)]
-    public async Task<ActionResult<Guid>> ExecuteAppCreation([FromBody] AppRequestModel appRequestModel)
-    {
-        var appId = await _appsBusinessLogic.AddAppAsync(appRequestModel).ConfigureAwait(false);
-        return CreatedAtRoute(nameof(GetAppDetailsByIdAsync), new {appId = appId}, appId);
-    }
-
-    /// <summary>
-    /// Retrieves all in review status apps in the marketplace .
-    /// </summary>
-    /// <param name="page">page index start from 0</param>
-    /// <param name="size">size to get number of records</param>
-    /// <returns>Collection of all in review status marketplace apps.</returns>
-    /// <remarks>Example: GET: /api/apps/inReview</remarks>
-    /// <response code="200">Returns the list of all in review status marketplace apps.</response>
-    [HttpGet]
-    [Route("inReview")]
-    [Authorize(Roles = "approve_app_release,decline_app_release")]
-    [ProducesResponseType(typeof(Pagination.Response<InReviewAppData>), StatusCodes.Status200OK)]
-    public Task<Pagination.Response<InReviewAppData>> GetAllInReviewStatusAppsAsync([FromQuery] int page = 0,[FromQuery] int size = 15) =>
-        _appsBusinessLogic.GetAllInReviewStatusAppsAsync( page,size);
-
-    /// <summary>
-    /// Submit an app for release
-    /// </summary>
-    /// <param name="appId" example="D3B1ECA2-6148-4008-9E6C-C1C2AEA5C645">ID of the app.</param>
-    /// <remarks>Example: PUT: /api/apps/D3B1ECA2-6148-4008-9E6C-C1C2AEA5C645/submit</remarks>
-    /// <response code="204">The app was successfully submitted for release.</response>
-    /// <response code="400">Either the sub claim is empty/invalid, user does not exist or the subscription might not have the correct status or the companyID is incorrect.</response>
-    /// <response code="404">App does not exist.</response>
-    [HttpPut]
-    [Route("{appId}/submit")]
-    [Authorize(Roles = "add_app")]
-    [ProducesResponseType(StatusCodes.Status204NoContent)]
-    [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
-    public async Task<IActionResult> SubmitAppReleaseRequest([FromRoute] Guid appId)
-    {
-        await this.WithIamUserId(userId => _appsBusinessLogic.SubmitAppReleaseRequestAsync(appId, userId)).ConfigureAwait(false);
-        return NoContent();
-    }
 
     /// <summary>
     /// Auto setup the app
